requirements.txt
streamlit
numpy
tensorflow
torch
pillow
groq
